chore: embed MCP server as hidden mcp-server subcommand
Move the stdio MCP server PoC into src/mcp/, drop the standalone mcp/
crate, and expose it via a hidden Command::McpServer so the claude
backend can self-spawn skill-forge over --mcp-config without shipping
a second binary. The handlers reuse host_call_llm and exec_cmd_impl
directly; submit_generated_code is a sink (extraction happens on the
claude side via stream-json).
