--- mcp/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "mcp-poc-server"
-version = "0.1.0"
-edition = "2024"
-
-[dependencies]
-serde = { version = "1", features = ["derive"] }
-serde_json = "1"
-reqwest = { version = "0.12", default-features = false, features = ["blocking", "rustls-tls"] }
-
-[[bin]]
-name = "mcp-poc-server"
-path = "src/main.rs"
-
-[workspace]
