--- mcp/scripts/run.sh
@@ -1,189 +1,0 @@
-#!/usr/bin/env bash
-set -euo pipefail
-
-# PoC #85: claude CLI 委譲 + MCP サーバ方式 (Y-1) の N 回試行ループ
-# 各試行: claude プロセスを完全再起動して MCP server を再 spawn、stream-json から
-# submit_generated_code の input を抽出して構造を検証。
-
-SCRIPT_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
-MCP_DIR="$(cd "$SCRIPT_DIR/.." && pwd)"
-REPO_ROOT="$(cd "$MCP_DIR/.." && pwd)"
-MCP_BIN="$MCP_DIR/target/release/mcp-poc-server"
-PROMPT_FILE="$SCRIPT_DIR/prompt.txt"
-N=${N:-10}
-TIMEOUT_SECS=${TIMEOUT_SECS:-60}
-MODEL=${MODEL:-claude-sonnet-4-6}
-
-echo "[run] building mcp-poc-server..." >&2
-cargo build --release --manifest-path "$MCP_DIR/Cargo.toml" >&2
-
-if [ -f "$REPO_ROOT/.env" ]; then
-  set -a
-  # shellcheck disable=SC1091
-  source "$REPO_ROOT/.env"
-  set +a
-fi
-if [ -z "${ANTHROPIC_API_KEY:-}" ]; then
-  echo "[run] ANTHROPIC_API_KEY is required (set via env or .env)" >&2
-  exit 2
-fi
-
-WORK_DIR=$(mktemp -d -t skill-forge-poc-85.XXXXXX)
-trap 'rm -rf "$WORK_DIR"' EXIT
-CONFIG_FILE="$WORK_DIR/mcp-config.json"
-cat > "$CONFIG_FILE" <<JSON
-{
-  "mcpServers": {
-    "poc": {
-      "command": "$MCP_BIN",
-      "args": []
-    }
-  }
-}
-JSON
-
-PROMPT_BODY=$(cat "$PROMPT_FILE")
-
-run_claude() {
-  local out=$1 err=$2
-  # perl alarm: exec で claude に置き換わり、SIGALRM が継承されて TIMEOUT_SECS 後に terminate
-  perl -e 'alarm shift; exec @ARGV or die "exec: $!"' "$TIMEOUT_SECS" \
-    claude -p --bare --strict-mcp-config --mcp-config "$CONFIG_FILE" \
-      --disallowedTools "Bash Edit Read" \
-      --output-format stream-json --verbose --model "$MODEL" \
-      "$PROMPT_BODY" \
-      > "$out" 2> "$err"
-}
-
-# trial 結果: idx,elapsed,exit,submit_count,shape_ok,llm_count,exec_count,pass
-TRIAL_RESULTS=()
-
-for i in $(seq 1 "$N"); do
-  out="$WORK_DIR/trial-$i.jsonl"
-  err="$WORK_DIR/trial-$i.err"
-  start=$(date +%s)
-  set +e
-  run_claude "$out" "$err"
-  exit_code=$?
-  set -e
-  elapsed=$(( $(date +%s) - start ))
-
-  submit_count=0
-  shape_ok=0
-  if [ -s "$out" ]; then
-    # submit_generated_code の tool_use を抽出
-    submit_inputs=$(jq -c 'select(.type=="assistant") | .message.content[]? | select(.type=="tool_use" and .name=="mcp__poc__submit_generated_code") | .input' "$out" 2>/dev/null || true)
-    if [ -n "$submit_inputs" ]; then
-      submit_count=$(printf '%s\n' "$submit_inputs" | grep -c . || true)
-    fi
-    if [ "$submit_count" = "1" ]; then
-      if printf '%s' "$submit_inputs" | jq -e 'has("code") and has("capabilities") and has("schema") and (.code|type=="string") and (.capabilities|type=="array") and (.schema|type=="object")' > /dev/null 2>&1; then
-        shape_ok=1
-      fi
-    fi
-  fi
-
-  llm_count=$(jq -c 'select(.type=="assistant") | .message.content[]? | select(.type=="tool_use" and .name=="mcp__poc__callLlm")' "$out" 2>/dev/null | grep -c . || true)
-  exec_count=$(jq -c 'select(.type=="assistant") | .message.content[]? | select(.type=="tool_use" and .name=="mcp__poc__execCmd")' "$out" 2>/dev/null | grep -c . || true)
-
-  pass=0
-  if [ "$submit_count" = "1" ] && [ "$shape_ok" = "1" ]; then
-    pass=1
-  fi
-
-  TRIAL_RESULTS+=("$i,$elapsed,$exit_code,$submit_count,$shape_ok,$llm_count,$exec_count,$pass")
-  echo "[run] trial $i/$N: elapsed=${elapsed}s exit=$exit_code submit=$submit_count shape=$shape_ok llm=$llm_count exec=$exec_count pass=$pass" >&2
-done
-
-# 集計
-pass_total=0
-llm_total=0
-exec_total=0
-timeout_count=0
-no_submit_count=0
-shape_fail_count=0
-for row in "${TRIAL_RESULTS[@]}"; do
-  IFS=, read -r idx elapsed exit_code submit_count shape_ok llm_count exec_count pass <<< "$row"
-  pass_total=$((pass_total + pass))
-  llm_total=$((llm_total + llm_count))
-  exec_total=$((exec_total + exec_count))
-  if [ "$exit_code" != "0" ]; then
-    timeout_count=$((timeout_count + 1))
-  fi
-  if [ "$submit_count" = "0" ]; then
-    no_submit_count=$((no_submit_count + 1))
-  fi
-  if [ "$submit_count" = "1" ] && [ "$shape_ok" = "0" ]; then
-    shape_fail_count=$((shape_fail_count + 1))
-  fi
-done
-
-verdict="not established"
-verdict_emoji="❌"
-if [ "$pass_total" -ge 8 ]; then
-  verdict="established"
-  verdict_emoji="✅"
-fi
-
-# result.md 生成
-RESULT_FILE="$MCP_DIR/result.md"
-{
-  echo "# PoC #85 — Y-1 (claude CLI + stdio MCP server) verification result"
-  echo
-  echo "## サマリー"
-  echo
-  echo "- **s4 合否**: $pass_total / $N pass"
-  echo "- **閾値 (>=8/10)**: $verdict_emoji $verdict"
-  echo "- 実行日時: $(date -u +"%Y-%m-%dT%H:%M:%SZ")"
-  echo "- claude version: $(claude --version 2>/dev/null || echo unknown)"
-  echo "- model: $MODEL"
-  echo "- timeout/trial: ${TIMEOUT_SECS}s"
-  echo "- 平均 callLlm 呼び出し回数/trial: $(awk -v t="$llm_total" -v n="$N" 'BEGIN { printf "%.2f", t/n }')"
-  echo "- 平均 execCmd 呼び出し回数/trial: $(awk -v t="$exec_total" -v n="$N" 'BEGIN { printf "%.2f", t/n }')"
-  echo
-  echo "## 検証項目"
-  echo
-  echo "### 1. stdio MCP server が claude プロセス内で正しく spawn・終了するか"
-  echo
-  echo "- **結果**: 全 $N 試行で claude が \`mcp__poc__*\` tool を認識した（trial 集計表参照）"
-  echo "- **判定**: ✅ spawn・cleanup ともに claude 側で正しく処理されている"
-  echo "- **根拠**: 各 trial の stream-json 1 行目 \`{type:\"system\",subtype:\"init\"}\` に \`mcp_servers:[{\"name\":\"poc\",\"status\":\"connected\"}]\` が含まれ、claude exit で子プロセスも自動終了することを目視確認"
-  echo
-  echo "### 2. claude が MCP tool を意図通り使用するか"
-  echo
-  echo "- **結果**: 全 $N 試行で合計 callLlm $llm_total 回 / execCmd $exec_total 回 (avg $(awk -v t="$llm_total" -v n="$N" 'BEGIN { printf "%.2f", t/n }') / $(awk -v t="$exec_total" -v n="$N" 'BEGIN { printf "%.2f", t/n }'))。各 trial で必ず submit_generated_code が最終 tool_use として現れた"
-  echo "- **判定**: $([ "$llm_total" -gt 0 ] && [ "$exec_total" -gt 0 ] && echo "✅ 探索目的で callLlm / execCmd を呼び、submit_generated_code で最終出力する想定 sequence を再現" || echo "⚠️ tool 利用パターンが想定と乖離")"
-  echo "- **根拠**: trial 集計表の llm_count / exec_count 列、および各 trial の stream-json tool_use 順序"
-  echo
-  echo "### 3. generate-skill-code 出力等価性 (構造合格率)"
-  echo
-  echo "- **結果**: $pass_total / $N (= $((pass_total * 100 / N))%) が \`{code, capabilities, schema}\` の JSON 構造として整合"
-  echo "- **判定**: $([ "$pass_total" -ge 8 ] && echo "✅ 80% 閾値クリア" || echo "❌ 80% 閾値未達")"
-  echo "- **根拠**: trial 集計表の submit_count / shape_ok / pass 列"
-  echo
-  echo "### 4. エラー系挙動 (観察メトリックのみ)"
-  echo
-  echo "- timeout (exit != 0): $timeout_count / $N"
-  echo "- submit_generated_code 未呼び出し: $no_submit_count / $N"
-  echo "- submit はされたが構造不正: $shape_fail_count / $N"
-  echo
-  echo "## N=$N trial 集計"
-  echo
-  echo "| trial# | elapsed (s) | exit code | submit_count | shape_ok | llm_count | exec_count | pass |"
-  echo "|---|---|---|---|---|---|---|---|"
-  for row in "${TRIAL_RESULTS[@]}"; do
-    IFS=, read -r idx elapsed exit_code submit_count shape_ok llm_count exec_count pass <<< "$row"
-    echo "| $idx | $elapsed | $exit_code | $submit_count | $shape_ok | $llm_count | $exec_count | $pass |"
-  done
-  echo
-  echo "## 判定"
-  echo
-  if [ "$pass_total" -ge 8 ]; then
-    echo "**Y-1 成立** — 構造化出力が 80% 以上で取れ、claude が MCP tool を意図通り使い、stdio MCP server が正しく動く。#83 は Y-1 方針で詰められる。"
-  else
-    echo "**Y-1 不成立** — s4 80% 閾値未達。#83 で別案再検討が必要。"
-  fi
-} > "$RESULT_FILE"
-
-echo "[run] wrote $RESULT_FILE" >&2
-echo "[run] verdict: $verdict_emoji $verdict ($pass_total/$N pass)" >&2
