+++ src/mcp/mod.rs
@@ -1,0 +1,4 @@
+mod server;
+mod tools;
+
+pub use server::run;
